Add OJS-XML worker. Also change XML-templating engine to Jinja2.
The XML generator worker is now able to do both marc-XML and custom XML
used by OJS-Import. The worker does some minor data preparation and
then fills the template document.

Also the loading of the JSON-data-file is now done within the actual
worker and not in the celery connector. This way it can be tested and
there is more separation between Celery and the actual worker.

The Tests for now only check for the existence of the generated files.
Tests for actual data integrity will follow.

Task #8639