Implement generic XML generator
Additionally there is another generator for marc-XML which calls the
generic one.

To make the XML generation tasks more generic, they now take
the XML path as a parameter. Also the data preparation logic is now
within the XML Jinja2 template.

The metadata JSON is now separated in a dedicated class, which is to be
extended in the future.