Remove object_id parameter from job handling
The object of the task chain will no longer be defined by the name of
the folder in the staging area. Instead, the files that should be
handled in a job will be given to retrieve_from_staging as a parameter.

Therefore there is no longer a need for providing an object_id when
creating a job. Depending on the job type the id can now be generated
or read from the metadata as part of the job chain.

Since retrieve_staging needs an id under which the result of the chain
is published it uses the job_id until proper mechanisms for reading
the object id are in place.

See #8718