Add job_type controller
It serves a JSON list containing the job types found in the config dir.
From the job files the meta information, called 'about', is read and served
as well.

The tests checks for the number of job types and if the first job type found
in the config dir is present.