Add build script and config files for auto generated API doc
The travis build script now also executes a script which generates the docu
and pushes the changes to a special branch "gh-pages". This branch contains
only the HTML documentation files and serves the documentation on the
internet. URL will be passed around later. Should be something like
'https://dai.github.io/cilantro/'.

For the docu generation all the python packages are imported, therefore it is
needed to have the libraries locally available, which is done via
pip requirements file.

All the needed files are generated by the script and removed afterward.
This way we only need a couple of config files. They are located in a new folder
'doc'.

The overall Look and Feel can surely be improved by using other templates
and sphinx extensions. This is only the first version.
Also a proper REST API docu will follow soon.