Extract pdf file names in split_pdf
The service method `split_merge_pdf()` should have no knowledge about
the workflow. Extracting the base name of the file in the method
assumes that the method knows that files are given as paths relative
to the staging directory and that create_object removes the folder
names.

This is extremely error prone since the behaviour could easily change
when the task chain in defined differently and is also not transparent
to the API user who might want to use the service outside the scope of
a specific workflow. Especially since this behaviour is not documented
anywhere.

Instead, extracting the base name is now done in the surrounding task.

This is still error prone to changes in the workflow (see #9090).