Add Publish-to-OJS worker
This worker publishes the documents referenced in the XML file found
in the working directory via OJS-Import-Plugin.

The Import-Plugin needs authentication which is (for now) hard-coded.

There is also a new test resource file, which is a simple text file
to be published via OJS. The OJS-Import-XMl file references are also
updated.

The unit test use the predefined XML from the test-resources.