Use representation name in foreach task
This allows a clearer distinction between the different partial results
as different representations.

This commit also tries to streamline the handling of paths in the
different convert tasks. As a first step all new files are added to
the origin representation. In the future the different tasks should
create more meaningful representations.

Also the handling of subobjects / parts-folders has been deactivated
temporarily. A new mechanism for explicitly targeting subobjects
has to be implemented (see #9130).

Closes #9126