Save task results as params
Return values of tasks are now saved under the 'result' key in the
params dictionary. This allows reading task results at a later stage,
i.e. in a following task or when querying the job status.

Results have to be dicts and are merged recursively so that partial
results in task chains accumulate and may be extended or modified
by following tasks.

See #9109