Recursively merge result lists
If groups are used inside a job chain (e.g. when using list_files or
list_parts) results are added to lists. In the case of nested list_-
tasks these lists are also nested. Therefore the result object is now
unpacked recursively and every list element is treated as a dict that
is merged with the result object.

Closes #9195