Split dev docker-compose config from base config
The config in docker-compose.override.yml is automatically merged
with the base config in docker-compose.yml. So the workflow for
running the containers in development stays the same. However this
allows explicitly using other overriding configs for different
environments such as production or testing.

See https://docs.docker.com/compose/production/