Add Generation of mets-XML
For DFG Viewer. Proper testing has to come, when files
are available online.

Also it is needed to have JPEG files generated and refenrenced in the XML,
because manual testing with DFG Viewer showed, that using TIFF does not
work, for reasons unknown. More testing and debugging has to be done.