#9412 Prevent overview from overflowing to fixate x-scrollbar
Using a new css-class "tableWrap", the main table in overview now has a fixed height and width, depending on the used screen, keeping the table from overflowing at the bottom and instead using a vertical scrollbar to navigate. The css-class "contentCol" gives every collumn of the main table a fixed minimum width, to prevent them from compressing when additional columns are displayed.