Extend docker-compose file for prod env
This is necessary since overriding docker-compose files in portainer
is not possible. So the production yml has to be standalone.

This also removes port mappings from the prod config. Since these
do not have to be exposed on the host. Instead a reverse proxy
service attached to the same docker network handles HTTP traffic
and routing.