Reset result attribute on every run in base task
This fixes a bug that caused older results from other jobs to be
returned in a seemingly random fashion.

As it turns out celery seems to reuse instances of task classes. This
lead to the fact that the results dictionary could already be populated
with values from previous runs on the same python thread.