Make sure previous results are passed to subtasks
Tasks that are created in list_files and list_parts now get an
additional param 'result' which is evaluated in addition to the
prev_result param in the base task.

This was necessary since results do not seem to be correctly
passed to child tasks by celery in every case.

See #9467