Add job database logic for creation and updating of jobs.
The database used is mongoDB which runs as a separate docker service.
The logic for interacting with the DB is encapsulated is a util module.

The job_type endpoint will now create a job in the job database.
This job can be updated by tasks belonging to the job.

The 'create_object' task sets the state to 'running', as this is
the first task of any job.
Additionally a new task 'finish_job' is introduced that sets the
state of the job to 'success'. It is supposed to be the last in
the task list.
In case of any task erroring out the base_task will catch any(!)
exception and set the job status to 'failed'.

The job controller gets a new endpoint to query all jobs of the logged
in user. They are delivered in JSON as list ob objects.