Add error handling for failed tasks.
The task is replaced by a special error-handler task, which writes
the failed state to the job database and ends any further task execution.

The problem of any exception thrown when handling task errors is solved
by introducing a special exception type 'ExceptionHandlingException'.

Error messages are a list in the job database. This makes it possible
to have tasks that throw an error but do not keep the chain from
continuing on. These messages may be collected at the end of the job.