Simplify docker-compose config
Since docker-compose pull is now used to explicitly pull images from
dockerhub and prevent building images in travis there is no need for
a separate docker-compose.ci.yml.

Also the contents of docker-compose.override.yml are now included in
docker-compose.yml as the distinction between dev and ci is not needed
anymore.