Add JSON-schema param validation.
The job-create endpoint now does some more checks on the existence
and validity of the POST parameters.
The parameters are passed on to job creation logic if the job definition
file and the related schema file are found and the validation against the
schema is passed.

The tests are updated accordingly and now also include checks for the
actual error message. Therefore a small test-schema is added for testing
validation.