Add ArachneDB-Mock-Container
This container is used for testing the export to the Arachne database.
This way the real database will not be affected by potentially bad
data used in cilantro-test runs.

There is a new switch for pointing cilantro to the real database.

The added SQL script is to set up some tables in the mock database to make
it look like the real database. It is taken by dumping the real one
without actual data.