New repository structure and endpoints
This includes a md5 checksum based structure of the repository, e.g. the object
"foo" is saved to "ac/bd/foo" instead of just "foo".
The endpoints now hide the actual structure of an object on the disk.
Instead it provides several routes to get (sub)objects, representations, files
of representations and files on root level (which normally contain metadata).