Fix file paths in METS XMLs
Before they pointed to the workspace directory, which is cleaned up
regularly. Correct is the repository dir.
The solution is a bit hacky. The path is split and put together, so the
part number can be kept, which would be rather complicated to get
otherwise.

Also the additional parameters for the template are extended to be able
to get environment paths.