Set fixed name for test container and use it for calling test commands
Since containers are dynamically named after the directory name
they are located in, there is no way to reliably know the name in advance.
This is needed though to call the right container by commands from outside.
It is used for calling the docu builder and calling the backend tests.