Add Jenkinsfile
This file is the config file for the Jenkins CI system. It contains
several steps: Starting docker containers, running backend tests,
running backend tests and building the documentation (for commits
on the master branch only). It uses the Makefile targets for this.
There is also a locking mechanism to make sure only one branch is built
in parallel, since there would be conflicts with the docker containers
otherwise.

In case of failure the docker logs are printed to the console.

In every build the logs are produced as artifact.