Merge METS-XML templates into one for all
This way code duplication is reduced. The control over what goes where
is done via task parameters.
'images_from_parts' means that sub-parts are iterated and images are
taken from there. Otherwise the image file are taken from the object
itself.
'is_subobject' helps for building the path to the file in the repository.
For subparts the parent object's ID is needed.