Fix staging response handling after successful upload
This only fixes the successful uploads (status code 200). The backend always receives a list of files, even if there was only one file selected by the user. If the list was received successfully, the backend sends a 200 status code, though it is possible that one of the files (or all) have an unknown/invalid extension. This change displays one list for all acknowledged files, as well as lists for rejected files (grouped by rejection reason).

Error handling for status codes other than 200 have still to be implemented!

See Redmine ticket #9681.