Add Zenon Import component
It is able to query Zenon and show the results as a table. The table 
includes buttons to take the ZenonID into the data fields and to go to 
the Zenon Page in a new tab.
Mising is the functionality to take over all data from Zenon into the 
form fields.