Extend FileBrowser
Include upload functionality and several capabilities to manipulate the 
file structure.
Since the FileBrowser can do all that, it is no longer needed in 
separate components.
Staging area now only consists of the FileBrowser, but is kept in case 
other stuff is needed there in the future.
The FileBrowser is also context aware as it consumes function provided 
by parent components. This way it can also be used in other contexts as 
the staging area.