Create Job names based on type and id
For jobs that start bulk jobs on multiple directories, we have multiple parameters that have different 'path' declarations. In order to generate the job name for the parent, we either would have to pick one or concatinate all existing paths. When working with a lot of directories, both solutions are may irritate the user.

So we generate the Job name based on type and ID for now.