Extend backend for hierarchical (batch) jobs
See Redmine ticket #11387.

In order to create a parent job that monitors multiple job chains running in parallel we use celery chords (See: https://docs.celeryproject.org/en/latest/userguide/canvas.html#chords). The task IDs for all subtasks have to be created before running the chord, in order to be able to keep references in the parent. We also need a new basic worker in order to react to the chords completion (i.e. writing the success to MongoDB). This additional worker is passed as a callback when creating the celery chord object.