Use /api as backend uri when no env var is set
This is necessary since the env var VUE_APP_BACKEND_URI is compiled into
the dist src when building the docker image and cannot be changed
afterwards. As a workaround the frontend assumes the relative uri '/api'
when VUE_APP_BACKEND_URI is not set. Therefore test and production
environments should be configured so that the backend is accessible
at the same domain as the frontend.

On development environments the env variable can still be overwritten
in .env files.