Fix AtoM API call to not crash when record is not found.
Since we use an endpoint on the backend instead of the native AtoM API 
we return { id: "not-found", message: "Information object not found" }
instead of a failed call. This means no exception is thrown. Therefore 
the old handler is never triggered. The error state is now checked in an 
if clause.