Add update functionality for the job list
To prevent having to refresh the page to see the latest status of jobs, 
they are updated automatically in the background. For that a list of 
pending jobs is generated and the status for the jobs is polled 
regularly and the job objects are updated accordingly. When there are no 
more pending jobs the polling is stopped.